test(logging): isolate real stream checks
